pyproject.toml
[project]
name = "academic_research"
version = "0.1"
description = "AI-driven agent designed to facilitate the exploration of the academic landscape"
authors = [{ name = "Antonio Gulli", email = "gulli@google.com" }]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^1.0.0"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
google-cloud-aiplatform = { version = "^1.93", extras = [
    "adk",
    "agent-engines",
] }

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { extras = ["eval"], version = "^1.0.0" }
google-cloud-aiplatform = { version = "^1.93", extras = [
    "adk",
    "agent-engines",
    "evaluation",
] }
pytest = "^8.3.5"
black = "^25.1.0"
pytest-asyncio = "^0.26.0"
pandas = "^2.2.3"
tabulate = "^0.9.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "asyncio: marks tests as async"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
